separate tags out to own object
Tags are a huge portion of our lookup information, so I'm hoping that by taking them out of the main Data Dictionary and to their own object (Data Tags) that they will make lookup of most things quicker.

This paves the way in case Prerequisites, Book Reference, and Notes needs to do the same.