extra covers, sheet display
extra covers: merit dots for more than 1 cover (due to primum 2+) are accounted for in the 'merits: power stat' calculation.

sheet display: the stretching display for 'cover (<name>): <value>' was giving the text width as '0' if there was no cover. now it default to the default text width display for the traits block.