starting xp/unspend
This version does not work, because it doesn't know the previous value to set to. It tries, but there is no easy way to discover it. Instead, we will be recording the old trait value from the freebie/spend. With new value, old value, and xp, we can revese everything without re-calculating anything.