add: old trait value and reason separation
The trait value that's getting overwritten by a freebie or spend is now stored properly in `trait_value_old`.

Because of this, reason longer needs to be stored as `stat 'from' -> 'to' for 'reason'` It is now simply stored as `reason`.

This will make an xp/unspend much easier, as we don't need to guess the old value.