bug: psychokinesis may only be taken once
The merit 'Psychokinesis' is in the book as a single merit. We gave it an instance/type in order to force people to choose a type when they took it, and to make it easier to see which type they have. Its prerequisite now prevents 'Psychokinesis' from having more than one instance/type.