add: allow multiple substats
&f.statpath-lookup-sheet: Mostly for the roller, but you may now name-complete "<trait>.<substat>.<substat>" as found on sheet.

In the nWoD and Chronicles of Darkness ("Storytelling") ruleset, you may roll multiple specialties on a skill, allowing for rolls of Crafts with both  the specialties of Clay and Pottery.

For the roller, this will be represeted as "crafts.clay.pottery". Our complex name-completion scheme will allow for "cra.cl.pot" and as long as 'cra' matches Crafts, the rest will follow.